+++ src/controllers/userController.js
@@ -1,103 +1,8 @@
 const Boom = require('@hapi/boom');
-const get = require('lodash/get');
 const translate = require('../helpers/translate');
 const UsersHelper = require('../helpers/users');
-const {
-  getUsersList,
-  getUsersListWithSectorHistories,
-  getLearnerList,
-  createAndSaveFile,
-  getUser,
-  userExists,
-} = require('../helpers/users');
 
 const { language } = translate;
-
-const authenticate = async (req, h) => {
-  try {
-    const authentication = await UsersHelper.authenticate(req.payload);
-    req.log('info', `${req.payload.email} connected`);
-
-    return h.response({
-      message: translate[language].userAuthentified,
-      data: { ...authentication },
-    }).state('alenvi_token', authentication.token)
-      .state('refresh_token', authentication.refreshToken)
-      .state('user_id', authentication.user._id);
-  } catch (e) {
-    req.log('error', e);
-    return Boom.isBoom(e) ? e : Boom.badImplementation(e);
-  }
-};
-
-/**
- * From MOBILE, refreshToken is in the payload
- * From WEBAPP, refresh_token is in the state
- */
-const refreshToken = async (req, h) => {
-  try {
-    const token = await UsersHelper.refreshToken(get(req, 'payload.refreshToken') || get(req, 'state.refresh_token'));
-
-    return h.response({
-      message: translate[language].userAuthentified,
-      data: { ...token },
-    }).state('alenvi_token', token.token)
-      .state('refresh_token', token.refreshToken)
-      .state('user_id', token.user._id);
-  } catch (e) {
-    req.log('error', e);
-    return Boom.isBoom(e) ? e : Boom.badImplementation(e);
-  }
-};
-
-const logout = async (req, h) => {
-  try {
-    return h.response({ message: translate[language].userAuthentified })
-      .unstate('alenvi_token')
-      .unstate('refresh_token')
-      .unstate('user_id');
-  } catch (e) {
-    req.log('error', e);
-    return Boom.isBoom(e) ? e : Boom.badImplementation(e);
-  }
-};
-
-const createPasswordToken = async (req) => {
-  try {
-    const passwordToken = await UsersHelper.createPasswordToken(req.payload.email);
-
-    return {
-      message: translate[language].resetPasswordTokenFound,
-      data: { passwordToken },
-    };
-  } catch (e) {
-    req.log('error', e);
-    return Boom.isBoom(e) ? e : Boom.badImplementation(e);
-  }
-};
-
-const forgotPassword = async (req) => {
-  try {
-    const mailInfo = await UsersHelper.forgotPassword(req.payload.email);
-
-    return { message: translate[language].emailSent, data: { mailInfo } };
-  } catch (e) {
-    req.log('error', e);
-    return Boom.isBoom(e) ? e : Boom.badImplementation(e);
-  }
-};
-
-const checkResetPasswordToken = async (req, h) => {
-  try {
-    const token = await UsersHelper.checkResetPasswordToken(req.params.token);
-
-    return h.response({ message: translate[language].resetPasswordTokenFound, data: { ...token } })
-      .state('alenvi_token', token.token);
-  } catch (e) {
-    req.log('error', e);
-    return Boom.isBoom(e) ? e : Boom.badImplementation(e);
-  }
-};
 
 const create = async (req) => {
   try {
@@ -120,7 +25,7 @@ const create = async (req) => {
 
 const list = async (req) => {
   try {
-    const users = await getUsersList(req.query, req.auth.credentials);
+    const users = await UsersHelper.getUsersList(req.query, req.auth.credentials);
 
     return {
       message: users.length === 0 ? translate[language].usersNotFound : translate[language].userFound,
@@ -134,7 +39,7 @@ const list = async (req) => {
 
 const listWithSectorHistories = async (req) => {
   try {
-    const users = await getUsersListWithSectorHistories(req.query, req.auth.credentials);
+    const users = await UsersHelper.getUsersListWithSectorHistories(req.query, req.auth.credentials);
 
     return {
       message: users.length === 0 ? translate[language].usersNotFound : translate[language].userFound,
@@ -148,7 +53,7 @@ const listWithSectorHistories = async (req) => {
 
 const activeList = async (req) => {
   try {
-    const users = await getUsersList(req.query, req.auth.credentials);
+    const users = await UsersHelper.getUsersList(req.query, req.auth.credentials);
     const activeUsers = users.filter(user => user.isActive);
 
     return {
@@ -163,7 +68,7 @@ const activeList = async (req) => {
 
 const learnerList = async (req) => {
   try {
-    const users = await getLearnerList(req.query, req.auth.credentials);
+    const users = await UsersHelper.getLearnerList(req.query, req.auth.credentials);
 
     return {
       message: users.length === 0 ? translate[language].usersNotFound : translate[language].userFound,
@@ -177,7 +82,7 @@ const learnerList = async (req) => {
 
 const show = async (req) => {
   try {
-    const user = await getUser(req.params._id, req.auth.credentials);
+    const user = await UsersHelper.getUser(req.params._id, req.auth.credentials);
 
     return {
       message: translate[language].userFound,
@@ -191,7 +96,7 @@ const show = async (req) => {
 
 const exists = async (req) => {
   try {
-    const userInfo = await userExists(req.query.email, req.auth.credentials);
+    const userInfo = await UsersHelper.userExists(req.query.email, req.auth.credentials);
 
     return {
       message: translate[language].userFound,
@@ -224,20 +129,6 @@ const update = async (req) => {
   }
 };
 
-const updatePassword = async (req) => {
-  try {
-    const updatedUser = await UsersHelper.updatePassword(req.params._id, req.payload);
-
-    return {
-      message: translate[language].userUpdated,
-      data: { updatedUser },
-    };
-  } catch (e) {
-    req.log('error', e);
-    return Boom.isBoom(e) ? e : Boom.badImplementation(e);
-  }
-};
-
 const updateCertificates = async (req) => {
   try {
     await UsersHelper.updateUserCertificates(req.params._id, req.payload, req.auth.credentials);
@@ -262,7 +153,7 @@ const removeHelper = async (req) => {
 
 const uploadFile = async (req) => {
   try {
-    const uploadedFile = await createAndSaveFile(req.params, req.payload);
+    const uploadedFile = await UsersHelper.createAndSaveFile(req.params, req.payload);
     return { message: translate[language].fileCreated, data: { uploadedFile } };
   } catch (e) {
     req.log('error', e);
@@ -312,10 +203,7 @@ const createDriveFolder = async (req) => {
 };
 
 module.exports = {
-  authenticate,
-  logout,
   create,
-  createPasswordToken,
   list,
   listWithSectorHistories,
   activeList,
@@ -324,13 +212,9 @@ module.exports = {
   exists,
   update,
   removeHelper,
-  refreshToken,
-  forgotPassword,
-  checkResetPasswordToken,
   updateCertificates,
   uploadFile,
   uploadPicture,
   deletePicture,
   createDriveFolder,
-  updatePassword,
 };
