+++ src/helpers/users.js
@@ -1,6 +1,5 @@
 const Boom = require('@hapi/boom');
 const moment = require('moment');
-const bcrypt = require('bcrypt');
 const pickBy = require('lodash/pickBy');
 const get = require('lodash/get');
 const has = require('lodash/has');
@@ -11,60 +10,15 @@ const { v4: uuidv4 } = require('uuid');
 const Role = require('../models/Role');
 const User = require('../models/User');
 const Company = require('../models/Company');
-const { TOKEN_EXPIRE_TIME } = require('../models/User');
 const Contract = require('../models/Contract');
 const translate = require('./translate');
 const GCloudStorageHelper = require('./gCloudStorage');
-const AuthenticationHelper = require('./authentication');
 const { TRAINER, AUXILIARY_ROLES, HELPER, AUXILIARY_WITHOUT_COMPANY, MOBILE } = require('./constants');
 const SectorHistoriesHelper = require('./sectorHistories');
-const EmailHelper = require('./email');
 const GdriveStorageHelper = require('./gdriveStorage');
 const UtilsHelper = require('./utils');
 
 const { language } = translate;
-
-exports.authenticate = async (payload) => {
-  const user = await User
-    .findOne({ 'local.email': payload.email.toLowerCase() })
-    .select('local refreshToken')
-    .lean();
-  const correctPassword = get(user, 'local.password') || '';
-  const isCorrect = await bcrypt.compare(payload.password, correctPassword);
-  if (!user || !user.refreshToken || !correctPassword || !isCorrect) throw Boom.unauthorized();
-
-  const tokenPayload = { _id: user._id.toHexString() };
-  const token = AuthenticationHelper.encode(tokenPayload, TOKEN_EXPIRE_TIME);
-
-  if (payload.origin === MOBILE && !user.firstMobileConnection) {
-    await User.updateOne(
-      { _id: user._id, firstMobileConnection: { $exists: false } },
-      { $set: { firstMobileConnection: moment().toDate() } }
-    );
-  }
-
-  return {
-    token,
-    tokenExpireDate: moment().add(TOKEN_EXPIRE_TIME, 'seconds').toDate(),
-    refreshToken: user.refreshToken,
-    user: tokenPayload,
-  };
-};
-
-exports.refreshToken = async (refreshToken) => {
-  const user = await User.findOne({ refreshToken }).lean();
-  if (!user) throw Boom.unauthorized();
-
-  const tokenPayload = { _id: user._id.toHexString() };
-  const token = AuthenticationHelper.encode(tokenPayload, TOKEN_EXPIRE_TIME);
-
-  return {
-    token,
-    tokenExpireDate: moment().add(TOKEN_EXPIRE_TIME, 'seconds').toDate(),
-    refreshToken,
-    user: tokenPayload,
-  };
-};
 
 exports.formatQueryForUsersList = async (query) => {
   const params = { ...pickBy(omit(query, ['role'])) };
@@ -263,12 +217,6 @@ exports.updateUser = async (userId, userPayload, credentials, canEditWithoutComp
   await User.updateOne(query, { $set: flat(payload) });
 };
 
-exports.updatePassword = async (userId, userPayload) => User.findOneAndUpdate(
-  { _id: userId },
-  { $set: flat(userPayload), $unset: { passwordToken: '' } },
-  { new: true }
-).lean();
-
 exports.updateUserCertificates = async (userId, userPayload, credentials) => {
   const companyId = get(credentials, 'company._id', null);
 
@@ -291,33 +239,6 @@ exports.updateUserInactivityDate = async (user, contractEndDate, credentials) =>
       { $set: { inactivityDate: moment(contractEndDate).add('1', 'month').startOf('M').toDate() } }
     );
   }
-};
-
-exports.checkResetPasswordToken = async (token) => {
-  const filter = { passwordToken: { token, expiresIn: { $gt: Date.now() } } };
-  const user = await User.findOne(flat(filter, { maxDepth: 2 })).select('local').lean();
-  if (!user) throw Boom.notFound(translate[language].userNotFound);
-
-  const payload = { _id: user._id, email: user.local.email };
-  const userPayload = pickBy(payload);
-  const expireTime = 86400;
-
-  return { token: AuthenticationHelper.encode(userPayload, expireTime), user: userPayload };
-};
-
-exports.createPasswordToken = async email => exports.generatePasswordToken(email, 24 * 3600 * 1000); // 1 day
-
-exports.forgotPassword = async (email) => {
-  const passwordToken = await exports.generatePasswordToken(email, 3600000);
-  return EmailHelper.forgotPasswordEmail(email, passwordToken);
-};
-
-exports.generatePasswordToken = async (email, time) => {
-  const payload = { passwordToken: { token: uuidv4(), expiresIn: Date.now() + time } };
-  const user = await User.findOneAndUpdate({ 'local.email': email }, { $set: payload }, { new: true }).lean();
-  if (!user) throw Boom.notFound(translate[language].userNotFound);
-
-  return payload.passwordToken;
 };
 
 exports.removeHelper = async (user) => {
