+++ src/routes/users.js
@@ -4,10 +4,7 @@ const Joi = require('joi');
 Joi.objectId = require('joi-objectid')(Joi);
 
 const {
-  authenticate,
-  logout,
   create,
-  createPasswordToken,
   list,
   listWithSectorHistories,
   activeList,
@@ -16,15 +13,11 @@ const {
   exists,
   update,
   removeHelper,
-  refreshToken,
-  forgotPassword,
-  checkResetPasswordToken,
   updateCertificates,
   uploadFile,
   uploadPicture,
   deletePicture,
   createDriveFolder,
-  updatePassword,
 } = require('../controllers/userController');
 const { CIVILITY_OPTIONS } = require('../models/schemaDefinitions/identity');
 const { ORIGIN_OPTIONS } = require('../models/User');
@@ -61,22 +54,6 @@ const driveUploadKeys = [
 exports.plugin = {
   name: 'routes-users',
   register: async (server) => {
-    server.route({
-      method: 'POST',
-      path: '/authenticate',
-      options: {
-        validate: {
-          payload: Joi.object().keys({
-            email: Joi.string().email().required(),
-            password: Joi.string().required(),
-            origin: Joi.string().valid(...ORIGIN_OPTIONS),
-          }).required(),
-        },
-        auth: false,
-      },
-      handler: authenticate,
-    });
-
     server.route({
       method: 'POST',
       path: '/',
@@ -314,41 +291,6 @@ exports.plugin = {
 
     server.route({
       method: 'PUT',
-      path: '/{_id}/create-password-token',
-      options: {
-        auth: { scope: ['users:edit', 'user:edit-{params._id}'] },
-        validate: {
-          params: Joi.object({ _id: Joi.objectId().required() }),
-          payload: Joi.object().keys({
-            email: Joi.string().email().required(),
-          }),
-        },
-        pre: [
-          { method: getUser, assign: 'user' },
-          { method: authorizeUserUpdate },
-        ],
-      },
-      handler: createPasswordToken,
-    });
-
-    server.route({
-      method: 'PUT',
-      path: '/{_id}/password',
-      options: {
-        auth: { scope: ['user:edit-{params._id}'] },
-        validate: {
-          params: Joi.object({ _id: Joi.objectId().required() }),
-          payload: Joi.object().keys({
-            local: Joi.object().keys({ password: Joi.string().min(6).required() }),
-            isConfirmed: Joi.boolean(),
-          }),
-        },
-      },
-      handler: updatePassword,
-    });
-
-    server.route({
-      method: 'PUT',
       path: '/{_id}/certificates',
       options: {
         auth: { scope: ['users:edit', 'user:edit-{params._id}'] },
@@ -380,49 +322,6 @@ exports.plugin = {
         ],
       },
       handler: removeHelper,
-    });
-
-    server.route({
-      method: 'POST',
-      path: '/refreshToken',
-      options: {
-        auth: false,
-        state: { parse: true, failAction: 'error' },
-      },
-      handler: refreshToken,
-    });
-
-    server.route({
-      method: 'POST',
-      path: '/logout',
-      options: { auth: false },
-      handler: logout,
-    });
-
-    server.route({
-      method: 'POST',
-      path: '/forgot-password',
-      options: {
-        validate: {
-          payload: Joi.object().keys({
-            email: Joi.string().email().required(),
-          }),
-        },
-        auth: false,
-      },
-      handler: forgotPassword,
-    });
-
-    server.route({
-      method: 'GET',
-      path: '/check-reset-password/{token}',
-      options: {
-        validate: {
-          params: Joi.object().keys({ token: Joi.string().required() }),
-        },
-        auth: false,
-      },
-      handler: checkResetPasswordToken,
     });
 
     server.route({
